Updated: API, Web and Documents
API:
- [Added] CartController (with add, update, remove and removeall).
- [Changed] Configured json export for UserCart.
- [Changed] UserCartService: Removed SubtractItemCount, added RemoveItem.

Web:
- [Changed] Cart count on bottom of user avatar on nav bar.
- [Added] Cart page (remove item, remove all).
- [Changed] ProductDetail: Add product to cart.
- [Added] ProductImage to get product preview image.
- [Added] CartRepository.

Document:
- [Changed] Draw.io for bumped database to version 0.6.1.

+++ src/api/Model/Product.cs
@@ -37,7 +37,7 @@ namespace PhoneStoreManager.Model
 
         [Required]
         [JsonProperty("price")]
-        public long Price { get; set; } = 0;
+        public double Price { get; set; } = 0;
 
         [JsonProperty("metadata")]
         public string? Metadata { get; set; } = null;
